requirements.txt
whisper
moviepy
pillow
numpy
requests
